# src/main/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.2.3-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `https://MY-REGISTRY-URL/apis/registry/v2` by default. Therefore you must prefix all API operation paths with `../apis/registry/v2` in this case. For example: `../apis/registry/v2/ids/globalIds/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
...
    },
    "paths": {
...
        "/system/info": {
            "summary": "Retrieve system information",
            "get": {
                "tags": [
                    "System"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SystemInfo"
                                }
                            }
                        },
                        "description": "On success, returns the system information."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getSystemInfo",
                "summary": "Get system information",
                "description": "This operation retrieves information about the running registry system, such as the version\nof the software and when it was built."
            }
        },
        "/search/artifacts": {
            "summary": "Search for artifacts in the registry.",
            "get": {
                "tags": [
                    "Search",
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "description": "Filter by artifact name.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "offset",
                        "description": "The number of artifacts to skip before starting to collect the result set.  Defaults to 0.",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "limit",
                        "description": "The number of artifacts to return.  Defaults to 20.",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "order",
                        "description": "Sort order, ascending (`asc`) or descending (`desc`).",
                        "schema": {
                            "$ref": "#/components/schemas/SortOrder"
                        },
                        "in": "query"
                    },
                    {
                        "name": "orderby",
                        "description": "The field to sort by.  Can be one of:\n\n* `name`\n* `createdOn`\n",
                        "schema": {
                            "$ref": "#/components/schemas/SortBy"
                        },
                        "in": "query"
                    },
                    {
                        "name": "labels",
                        "description": "Filter by label.  Include one or more label to only return artifacts containing all of the\nspecified labels.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "in": "query"
                    },
                    {
                        "name": "properties",
                        "description": "Filter by one or more name/value property.  Separate each name/value pair using a colon.  For\nexample `properties=foo:bar` will return only artifacts with a custom property named `foo`\nand value `bar`.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "in": "query"
                    },
                    {
                        "name": "description",
                        "description": "Filter by description.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "group",
                        "description": "Filter by artifact group.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "globalId",
                        "description": "Filter by globalId.",
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        },
                        "in": "query"
                    },
                    {
                        "name": "contentId",
                        "description": "Filter by contentId.",
                        "schema": {
                            "format": "int64",
                            "type": "integer"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArtifactSearchResults"
                                }
                            }
                        },
                        "description": "On a successful response, returns a result set of artifacts - one for each artifact\nin the registry that matches the criteria."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "searchArtifacts",
                "summary": "Search for artifacts",
                "description": "Returns a paginated list of all artifacts that match the provided filter criteria.\n"
            },
            "post": {
                "requestBody": {
                    "description": "The content to search for.",
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/File"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Search",
                    "Artifacts"
                ],
                "parameters": [
                    {
                        "name": "canonical",
                        "description": "Parameter that can be set to `true` to indicate that the server should \"canonicalize\" the content when searching for matching artifacts.  Canonicalization is unique to each artifact type, but typically involves removing any extra whitespace and formatting the content in a consistent manner.  Must be used along with the `artifactType` query parameter.",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query"
                    },
                    {
                        "name": "artifactType",
                        "description": "Indicates the type of artifact represented by the content being used for the search.  This is only needed when using the `canonical` query parameter, so that the server knows how to canonicalize the content prior to searching for matching artifacts.",
                        "schema": {
                            "$ref": "#/components/schemas/ArtifactType"
                        },
                        "in": "query"
                    },
                    {
                        "name": "offset",
                        "description": "The number of artifacts to skip before starting to collect the result set.  Defaults to 0.",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "limit",
                        "description": "The number of artifacts to return.  Defaults to 20.",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "order",
                        "description": "Sort order, ascending (`asc`) or descending (`desc`).",
                        "schema": {
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    },
                    {
                        "name": "orderby",
                        "description": "The field to sort by.  Can be one of:\n\n* `name`\n* `createdOn`\n",
                        "schema": {
                            "enum": [
                                "name",
                                "createdOn"
                            ],
                            "type": "string"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArtifactSearchResults"
                                }
                            }
                        },
                        "description": "On a successful response, returns a result set of artifacts - one for each artifact\nin the registry that matches the criteria."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "searchArtifactsByContent",
                "summary": "Search for artifacts by content",
                "description": "Returns a paginated list of all artifacts with at least one version that matches the\nposted content.\n"
            }
        },
...
        "/system/limits": {
            "summary": "Retrieve resource limits information",
            "get": {
                "tags": [
                    "System"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Limits"
                                }
                            }
                        },
                        "description": "On success, returns resource limits"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getResourceLimits",
                "summary": "Get resource limits information",
                "description": "This operation retrieves the list of limitations on used resources, that are applied on the current instance of Registry."
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Root Type for Error",
                "description": "All error responses, whether `4xx` or `5xx` will include one of these as the response\nbody.",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "The short error message.",
                        "type": "string"
                    },
                    "error_code": {
                        "format": "int32",
                        "description": "The server-side error code.",
                        "type": "integer"
                    },
                    "detail": {
                        "description": "Full details about the error.  This might contain a server stack trace, for example.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The error name - typically the classname of the exception thrown by the server.",
                        "type": "string"
                    }
                },
                "example": {
                    "error_code": 500,
                    "message": "An error occurred somewhere."
                }
            },
...
            "ArtifactType": {
                "description": "",
                "enum": [
                    "AVRO",
                    "PROTOBUF",
                    "JSON",
                    "OPENAPI",
                    "ASYNCAPI",
                    "GRAPHQL",
                    "KCONNECT",
                    "WSDL",
                    "XSD",
                    "XML"
                ],
                "type": "string",
                "example": "AVRO",
                "x-codegen-package": "io.apicurio.registry.types"
            },
...
            "ArtifactState": {
                "description": "Describes the state of an artifact or artifact version.  The following states\nare possible:\n\n* ENABLED\n* DISABLED\n* DEPRECATED\n",
                "enum": [
                    "ENABLED",
                    "DISABLED",
                    "DEPRECATED"
                ],
                "type": "string",
                "x-codegen-package": "io.apicurio.registry.types"
            },
            "ArtifactSearchResults": {
                "description": "Describes the response received when searching for artifacts.",
                "required": [
                    "count",
                    "artifacts"
                ],
                "type": "object",
                "properties": {
                    "artifacts": {
                        "description": "The artifacts returned in the result set.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SearchedArtifact"
                        }
                    },
                    "count": {
                        "description": "The total number of artifacts that matched the query that produced the result set (may be \nmore than the number of artifacts in the result set).",
                        "type": "integer"
                    }
                }
            },
            "SortOrder": {
                "description": "",
                "enum": [
                    "asc",
                    "desc"
                ],
                "type": "string"
            },
...
            "GroupId": {
                "description": "An ID of a single artifact group.",
                "type": "string",
                "example": "\"my-group\""
            },
            "ArtifactId": {
                "description": "The ID of a single artifact.",
                "type": "string",
                "example": "\"example-artifact\""
            },
...
            "SortBy": {
                "description": "",
                "enum": [
                    "name",
                    "createdOn"
                ],
                "type": "string"
            },
...
            "SystemInfo": {
                "title": "Root Type for SystemInfo",
                "description": "",
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    },
                    "builtOn": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "example": {
                    "name": "Apicurio Registry (SQL)",
                    "description": "The Apicurio Registry application.",
                    "version": "2.0.0.Final",
                    "builtOn": "2021-03-19T12:55:00Z"
                }
            },
...
            "SearchedArtifact": {
                "description": "Models a single artifact from the result set returned when searching for artifacts.",
                "required": [
                    "createdBy",
                    "createdOn",
                    "id",
                    "type",
                    "state",
                    "group"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/ArtifactId",
                        "description": ""
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "description": {
                        "description": "",
                        "type": "string"
                    },
                    "createdOn": {
                        "format": "date-time",
                        "description": "",
                        "type": "string",
                        "x-codegen-formatPattern": "yyyy-MM-dd'T'HH:mm:ssZ"
                    },
                    "createdBy": {
                        "description": "",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ArtifactType",
                        "description": ""
                    },
                    "labels": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "state": {
                        "$ref": "#/components/schemas/ArtifactState",
                        "description": ""
                    },
                    "modifiedOn": {
                        "format": "date-time",
                        "description": "",
                        "type": "string",
                        "x-codegen-formatPattern": "yyyy-MM-dd'T'HH:mm:ssZ"
                    },
                    "modifiedBy": {
                        "description": "",
                        "type": "string"
                    },
                    "groupId": {
                        "$ref": "#/components/schemas/GroupId",
                        "description": ""
                    }
                },
                "example": {
                    "groupId": "My-Group",
                    "id": "Procurement-Invoice",
                    "name": "Artifact Name",
                    "description": "Description of the artifact",
                    "labels": [
                        "current",
                        "internal"
                    ],
                    "type": "AVRO",
                    "state": "ENABLED",
                    "createdBy": "user1",
                    "createdOn": "2019-03-22T12:51:19Z"
                }
            },
...
            "Limits": {
                "title": "Root Type for Limits",
                "description": "List of limitations on used resources, that are applied on the current instance of Registry.\nKeys represent the resource type and are suffixed by the corresponding unit.\nValues are integers. Only non-negative values are allowed, with the exception of -1, which means that the limit is not applied.",
                "type": "object",
                "properties": {
                    "maxTotalSchemasCount": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxSchemaSizeBytes": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxArtifactsCount": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxVersionsPerArtifactCount": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxArtifactPropertiesCount": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxPropertyKeySizeBytes": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxPropertyValueSizeBytes": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxArtifactLabelsCount": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxLabelSizeBytes": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxNameLengthChars": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxDescriptionLengthChars": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "maxRequestsPerSecondCount": {
                        "format": "int64",
                        "type": "integer"
                    }
                },
                "example": {
                    "maxTotalSchemasCount": -1,
                    "maxSchemaSizeBytes": -1,
                    "maxArtifactsCount": -1,
                    "maxVersionsPerArtifactCount": -1,
                    "maxArtifactPropertiesCount": -1,
                    "maxPropertyKeySizeBytes": -1,
                    "maxPropertyValueSizeBytes": -1,
                    "maxArtifactLabelsCount": -1,
                    "maxLabelSizeBytes": -1,
                    "maxNameLengthChars": -1,
                    "maxDescriptionLengthChars": -1,
                    "maxRequestsPerSecondCount": -1
                }
            },
            "File": {
                "format": "binary",
                "type": "string",
                "x-codegen-inline": true
            }
        },
        "responses": {
...
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
...
        }
    },
...
}